+++ CHANGELOG.md
@@ -13,10 +13,11 @@ Possible log types:
 
 ### Unreleased
 
-- [added] Add Nordic's nRFx support
+- [added] Add OpenOCD debug probe interface
+- [added] Add Nordic's nRFx family support
+- [added] Add STM32F4 family support
 - [added] Add `linker.include` section to Drone.toml
 - [fixed] Fix `clippy::missing_safety_doc` lint for generated projects
-- [added] Add STM32F4 family support
 
 ### v0.10.1 (2019-09-27)